Moved file for experiments solely for the sake of learning to separate dir
I'm always going to be experimenting in source code in the
process of learning Agda, but some source files are just for
learning--they're not also an attempt to develop source for the
project. I'm putting files like that in agda/learning.  The
learning process might be in service of the project, but the code
in these files are not intended as possible candidates for
project code.