Completed using Vecs for consistent environment/dunlin configuration.
To create the dunlins in an env, we need a dunlin id and a dunlin
constructor for each, so with separate sequences of those ids
and constructors, it's worth making sure they are the same length.
Now they are Vecs instead of Lists, and an initial fst of a dep
pair binds their common lengths.